左侧规定比列加仓义务

+++ ai-invest-service/src/main/java/com/swak/ai/invest/service/data/StockQuote.java
@@ -2,6 +2,8 @@ package com.swak.ai.invest.service.data;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * @author: ljq
  * @date: 2024/11/23
@@ -12,10 +14,10 @@ public class StockQuote {
     private String tsCode;
 
     // 当前估计
-    private double currentPrice;
+    private BigDecimal currentPrice;
 
     // 涨跌幅
-    private double chg;
+    private BigDecimal chg;
 
 
 }
